fix: Changed the code to do the following 1. handle the int for balances instead of float or double 2. Made the hdfs generic such that any thing can be logged

+++ LedgerProxy/types/security.go
@@ -3,11 +3,11 @@ package types
 import "time"
 
 type SecureMessage struct {
-	Status bool        `json:"status"`
+	Status    bool      `json:"status"`
 	Timestamp time.Time `json:"timestamp"`
 	From      string    `json:"from"`
 	To        string    `json:"to"`
-	Amount    float64   `json:"amount"`
+	Amount    int64   `json:"amount"`
 	Message   string    `json:"message"`
 	Nonce     string    `json:"nonce"`
 	Hash      []byte    `json:"hash"`
@@ -15,37 +15,37 @@ type SecureMessage struct {
 
 type SecureOfflineWithdrawMessage struct {
 	AccountId string    `json:"account_id"`
-	Amount    float64   `json:"amount"`
+	Amount    int64   `json:"amount"`
 	Nonce     string    `json:"nonce"`
 	Timestamp time.Time `json:"timestamp"`
 }
 type SecureOfflineDepositMessage struct {
 	AccountId string    `json:"account_id"`
-	Amount    float64   `json:"amount"`
+	Amount    int64   `json:"amount"`
 	Nonce     string    `json:"nonce"`
 	Timestamp time.Time `json:"timestamp"`
 }
 
 type SecureCreateAccountMessage struct {
-	AccountId string    `json:"account_id"`
-	Balance   float64   `json:"balance"`
-	Nonce     string    `json:"nonce"`
+	AccountId string  `json:"account_id"`
+	Balance   int64 `json:"balance"`
+	Nonce     string  `json:"nonce"`
 }
 
 type SecureSyncMessage struct {
 	From      string    `json:"from"`
 	To        string    `json:"to"`
-	Amount    float64   `json:"amount"`
+	Amount    int64   `json:"amount"`
 	Nonce     string    `json:"nonce"`
 	Timestamp time.Time `json:"timestamp"`
 }
 
 type LedgerSyncMessage struct {
-	FromId      string    `json:"from_id"`
-	ToId        string    `json:"to_id"`
-	Amount    float64   `json:"amount"`
-	Nonce     string    `json:"nonce"`
-	OfflineTransaction bool `json:"offine_transaction"`
+	FromId             string  `json:"from_id"`
+	ToId               string  `json:"to_id"`
+	Amount             int64 `json:"amount"`
+	Nonce              string  `json:"nonce"`
+	OfflineTransaction bool    `json:"offine_transaction"`
 }
 
 type UnpackedMessage struct {
